:recycle: refactor: 구매자 시점 메이더 홈 url 변경 (#89)
