+++ src/main/java/umc/unimade/domain/accounts/service/BuyerQueryService.java
@@ -1,11 +1,15 @@
 package umc.unimade.domain.accounts.service;
 import lombok.*;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import umc.unimade.domain.accounts.dto.SellerPageResponse;
 import umc.unimade.domain.accounts.entity.Seller;
+import umc.unimade.domain.accounts.exception.SellerExceptionHandler;
+import umc.unimade.domain.accounts.repository.SellerRepository;
 import umc.unimade.domain.orders.dto.BuyerOrderHistoryResponse;
 import umc.unimade.domain.accounts.dto.BuyerPageResponse;
 import umc.unimade.domain.accounts.entity.Buyer;
@@ -23,6 +27,9 @@ import umc.unimade.domain.orders.entity.Orders;
 import umc.unimade.domain.orders.exception.OrderExceptionHandler;
 import umc.unimade.domain.orders.repository.OrderRepository;
 import umc.unimade.domain.products.entity.PickupOption;
+import umc.unimade.domain.products.entity.ProductStatus;
+import umc.unimade.domain.products.entity.Products;
+import umc.unimade.domain.products.repository.ProductRepository;
 import umc.unimade.global.common.ErrorCode;
 import umc.unimade.domain.accounts.exception.UserExceptionHandler;
 
@@ -37,6 +44,8 @@ public class BuyerQueryService {
     private final FavoriteSellerRepository favoriteSellerRepository;
     private final FavoriteProductRepository favoriteProductRepository;
     private final OrderRepository orderRepository;
+    private final SellerRepository sellerRepository;
+    private final ProductRepository productRepository;
 
     // 구매자 마이페이지
     public BuyerPageResponse getBuyerPage(Buyer buyer){
@@ -122,5 +131,31 @@ public class BuyerQueryService {
     private Orders findOrderById(Long orderId){
         return orderRepository.findById(orderId)
                 .orElseThrow(() -> new OrderExceptionHandler(ErrorCode.ORDER_NOT_FOUND));
+    }
+
+    // 구매자 시점 메이더 홈 - 특정 sellerId의 프로필 정보와 selling 상태인 상품 목록을 조회
+    public SellerPageResponse getSellerPage(Long sellerId, String sort, Pageable pageable) {
+        Seller seller = sellerRepository.findById(sellerId)
+                .orElseThrow(() -> new SellerExceptionHandler(ErrorCode.SELLER_NOT_FOUND));
+
+        // 정렬
+        Page<Products> products;
+        switch (sort) {
+            case "popular": // 인기순
+                products = productRepository.findBySellerIdAndStatusOrderByPopularity(sellerId, ProductStatus.SELLING, pageable);
+                break;
+            case "latest": // 최신순
+                products = productRepository.findBySellerIdAndStatusOrderByCreatedAtDesc(sellerId, ProductStatus.SELLING, pageable);
+                break;
+            case "deadline": // 마감순
+                products = productRepository.findBySellerIdAndStatusOrderByDeadline(sellerId, ProductStatus.SELLING, pageable);
+                break;
+            default:
+                products = productRepository.findBySellerIdAndStatusOrderByPopularity(sellerId, ProductStatus.SELLING, pageable);
+        }
+
+        Page<SellerPageResponse.ProductsResponse> productResponses = products.map(SellerPageResponse.ProductsResponse::from);
+
+        return SellerPageResponse.of(seller, productResponses);
     }
 }
